pyproject.toml
[project]
name = "telegram-chatbot-with-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.24",
    "langchain-core>=0.3.56",
    "langchain-google-genai>=2.1.3",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.0",
    "tgcrypto>=1.2.5",
]
